pyproject.toml
[tool.poetry]
name = "modern-maverick"
version = "0.1.0"
description = ""
authors = ["octavioflo <41705533+octavioflo@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "modern_maverick"}]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.10.2"
PyYAML = "^6.0"
python-nmap = "^0.7.1"
termcolor = "^1.1.0"
rich = "^12.6.0"


[tool.poetry.group.dev.dependencies]
black = "^22.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
